requirements.txt
sqlalchemy
fastapi
uvicorn
sqlalchemy
python-jose[cryptography] 
passlib[bcrypt] 
fastapi[all]
